Move salt.utils.arg_lookup to salt.utils.args.py
