+++ tests/unit/utils/test_args.py
@@ -5,10 +5,17 @@ from __future__ import absolute_import
 from collections import namedtuple
 
 # Import Salt Libs
+from salt.exceptions import SaltInvocationError
 import salt.utils.args
 
 # Import Salt Testing Libs
-from tests.support.unit import TestCase
+from tests.support.unit import TestCase, skipIf
+from tests.support.mock import (
+    DEFAULT,
+    NO_MOCK,
+    NO_MOCK_REASON,
+    patch
+)
 
 
 class ArgsTestCase(TestCase):
@@ -45,3 +52,33 @@ class ArgsTestCase(TestCase):
 
         ret = salt.utils.args.parse_kwarg('foobar')
         self.assertEqual(ret, (None, None))
+
+    def test_arg_lookup(self):
+        def dummy_func(first, second, third, fourth='fifth'):
+            pass
+
+        expected_dict = {'args': ['first', 'second', 'third'], 'kwargs': {'fourth': 'fifth'}}
+        ret = salt.utils.args.arg_lookup(dummy_func)
+        self.assertEqual(expected_dict, ret)
+
+    @skipIf(NO_MOCK, NO_MOCK_REASON)
+    def test_format_call(self):
+        with patch('salt.utils.args.arg_lookup') as arg_lookup:
+            def dummy_func(first=None, second=None, third=None):
+                pass
+
+            arg_lookup.return_value = {'args': ['first', 'second', 'third'], 'kwargs': {}}
+            get_function_argspec = DEFAULT
+            get_function_argspec.return_value = namedtuple('ArgSpec', 'args varargs keywords defaults')(
+                args=['first', 'second', 'third', 'fourth'], varargs=None, keywords=None, defaults=('fifth',))
+
+            # Make sure we raise an error if we don't pass in the requisite number of arguments
+            self.assertRaises(SaltInvocationError, salt.utils.format_call, dummy_func, {'1': 2})
+
+            # Make sure we warn on invalid kwargs
+            ret = salt.utils.format_call(dummy_func, {'first': 2, 'second': 2, 'third': 3})
+            self.assertGreaterEqual(len(ret['warnings']), 1)
+
+            ret = salt.utils.format_call(dummy_func, {'first': 2, 'second': 2, 'third': 3},
+                                         expected_extra_kws=('first', 'second', 'third'))
+            self.assertDictEqual(ret, {'args': [], 'kwargs': {}})
